+++ spec/iso8601/time_interval_spec.rb
@@ -54,19 +54,13 @@ RSpec.describe ISO8601::TimeInterval do
     context "allowed patterns" do
       it "should parse <start>/<duration>" do
         expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P1Y') }.to_not raise_error
-        expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P0.5Y') }.to_not raise_error
-        expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P1Y0.5M') }.to_not raise_error
-        expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P1Y0,5M') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P1Y1M1D') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P1Y1M1DT1H1M1.0S') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('2007-03-01T13:00:00Z/P1Y1M1DT1H1M1,0S') }.to_not raise_error
       end
 
       it "should parse <duration>/<end>" do
-        expect { ISO8601::TimeInterval.parse('P1Y0,5M/2010-05-09T10:30:12+04') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('P1Y1M1D/2010-05-09T10:30:12+04:00') }.to_not raise_error
-        expect { ISO8601::TimeInterval.parse('P1Y1M0.5D/2010-05-09T10:30:12-04:00') }.to_not raise_error
-        expect { ISO8601::TimeInterval.parse('P1Y1M0,5D/2010-05-09T10:30:12-00:00') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('P1Y1M1DT1H/-2014-05-31T16:26:00Z') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('P1Y1M1DT0.5H/2014-05-31T16:26:10.5Z') }.to_not raise_error
         expect { ISO8601::TimeInterval.parse('P1Y1M1DT0,5H/2014-05-31T16:26:10,5Z') }.to_not raise_error
@@ -178,6 +172,107 @@ RSpec.describe ISO8601::TimeInterval do
       expect(ISO8601::TimeInterval.parse(pattern2).first).to eq(start_time)
       expect(ISO8601::TimeInterval.parse(pattern3).first).to eq(start_time)
     end
+
+    describe "November" do
+      pairs = [
+        {pattern: 'P1Y/2017-11-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-11-09T07:00:00Z')},
+        {pattern: 'P1M/2017-11-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-10-09T07:00:00Z')},
+        {pattern: 'P1D/2017-11-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-11-08T07:00:00Z')},
+        {pattern: 'PT1H/2017-11-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-11-09T06:00:00Z')},
+      ]
+
+      pairs.each do |pair|
+        it "should calculate correctly the start_time for #{pair[:pattern]}" do
+          expect(ISO8601::TimeInterval.parse(pair[:pattern]).first.to_s).to eq(pair[:start_time].to_s)
+        end
+      end
+    end
+
+    describe "December" do
+      pairs = [
+        {pattern: 'P1Y/2017-12-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-12-09T07:00:00Z')},
+        {pattern: 'P1M/2017-12-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-11-09T07:00:00Z')},
+        {pattern: 'P3D/2017-12-06T18:30:00Z',
+         start_time: ISO8601::DateTime.new('2017-12-03T18:30:00Z')},
+        {pattern: 'P1D/2017-12-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-12-08T07:00:00Z')},
+        {pattern: 'PT1H/2017-12-09T07:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-12-09T06:00:00Z')},
+      ]
+
+      pairs.each do |pair|
+        it "should calculate correctly the start_time for #{pair[:pattern]}" do
+          expect(ISO8601::TimeInterval.parse(pair[:pattern]).first.to_s).to eq(pair[:start_time].to_s)
+        end
+      end
+    end
+
+    describe "January" do
+      pairs = [
+        {pattern: 'P1Y/2017-01-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-01-01T00:00:00Z')},
+        {pattern: 'P1M/2017-01-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-12-01T00:00:00Z')},
+        {pattern: 'P1D/2017-01-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-12-31T00:00:00Z')},
+        {pattern: 'PT1H/2017-01-01T01:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-01-01T00:00:00Z')},
+      ]
+
+      pairs.each do |pair|
+        it "should calculate correctly the start_time for #{pair[:pattern]}" do
+          expect(ISO8601::TimeInterval.parse(pair[:pattern]).first.to_s).to eq(pair[:start_time].to_s)
+        end
+      end
+    end
+
+    describe "February" do
+      pairs = [
+        {pattern: 'P1Y/2017-02-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-02-01T00:00:00Z')},
+        {pattern: 'P1M/2017-02-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-01-01T00:00:00Z')},
+        {pattern: 'P1D/2017-02-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-01-31T00:00:00Z')},
+        {pattern: 'PT1H/2017-02-01T01:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-02-01T00:00:00Z')},
+      ]
+
+      pairs.each do |pair|
+        it "should calculate correctly the start_time for #{pair[:pattern]}" do
+          expect(ISO8601::TimeInterval.parse(pair[:pattern]).first.to_s).to eq(pair[:start_time].to_s)
+        end
+      end
+    end
+
+    describe "March" do
+      pairs = [
+        {pattern: 'P1Y/2017-03-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2016-03-01T00:00:00Z')},
+        {pattern: 'P1M/2017-03-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-02-01T00:00:00Z')},
+        {pattern: 'P1D/2017-03-01T00:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-02-28T00:00:00Z')},
+        {pattern: 'PT1H/2017-03-01T01:00:00Z',
+         start_time: ISO8601::DateTime.new('2017-03-01T00:00:00Z')},
+      ]
+
+      pairs.each do |pair|
+        it "should calculate correctly the start_time for #{pair[:pattern]}" do
+          expect(ISO8601::TimeInterval.parse(pair[:pattern]).first.to_s).to eq(pair[:start_time].to_s)
+        end
+      end
+    end
+
+
+
+
   end
 
   describe "#last" do
@@ -205,8 +300,8 @@ RSpec.describe ISO8601::TimeInterval do
 
   describe "#to_s" do
     it "should return the pattern if TimeInterval is initialized with a pattern" do
-      pattern = 'P1Y1M1DT0,5H/2014-05-31T16:26:10,5Z'
-      pattern2 = '2007-03-01T13:00:00Z/P1Y0,5M'
+      pattern = 'P1Y1M1DT0,5S/2014-05-31T16:26:10Z'
+      pattern2 = '2007-03-01T13:00:00Z/P1Y'
 
       expect(ISO8601::TimeInterval.parse(pattern).to_s).to eq(pattern)
       expect(ISO8601::TimeInterval.parse(pattern2).to_s).to eq(pattern2)